Use real `CalculateSoundPosition` with NOSOUND
Fixes big-endian tests, which currently build with `-DNOSOUND`
to reduce build time.
