+++ Source/effects.cpp
@@ -8,6 +8,7 @@
 #include "engine/random.hpp"
 #include "engine/sound.h"
 #include "engine/sound_defs.hpp"
+#include "engine/sound_position.hpp"
 #include "init.h"
 #include "player.h"
 #include "utils/stdcompat/algorithm.hpp"
@@ -1199,25 +1200,6 @@ void stream_stop()
 		sgpStreamSFX->pSnd = nullptr;
 		sgpStreamSFX = nullptr;
 	}
-}
-
-bool CalculateSoundPosition(Point soundPosition, int *plVolume, int *plPan)
-{
-	const auto &playerPosition = MyPlayer->position.tile;
-	const auto delta = soundPosition - playerPosition;
-
-	int pan = (delta.deltaX - delta.deltaY) * 256;
-	*plPan = clamp(pan, PAN_MIN, PAN_MAX);
-
-	int volume = playerPosition.ApproxDistance(soundPosition);
-	volume *= -64;
-
-	if (volume <= ATTENUATION_MIN)
-		return false;
-
-	*plVolume = volume;
-
-	return true;
 }
 
 void PlaySFX(_sfx_id psfx)
