+++ Source/engine/sound_position.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "engine/point.hpp"
+
+namespace devilution {
+
+bool CalculateSoundPosition(Point soundPosition, int *plVolume, int *plPan);
+
+} // namespace devilution
